move ScriptEngine and related methods into own module
